# docs/swagger.yaml
definitions:
...
  ds.Banknote:
    properties:
      banknote_id:
        type: integer
      currency:
        type: string
      description:
        type: string
      image_url:
        type: string
      nominal:
        type: number
      status:
        type: string
    type: object
...
  ds.NewStatus:
    properties:
      operation_id:
        type: integer
      status:
        type: string
    type: object
  ds.Operation:
    properties:
      comletion_at:
        type: string
      crated_at:
        type: string
      formation_at:
        type: string
      is_income:
        type: string
      moderator:
        $ref: '#/definitions/ds.User'
      moderator_id:
        type: integer
      operation_banknote:
        items:
          $ref: '#/definitions/ds.OperationBanknote'
        type: array
      operation_id:
        type: integer
      status:
        type: string
      status_check:
        type: string
      user:
        $ref: '#/definitions/ds.User'
      user_id:
        type: integer
    type: object
  ds.OperationBanknote:
    properties:
      banknote:
        $ref: '#/definitions/ds.Banknote'
      banknote_id:
        type: integer
      id:
        type: integer
      operation_id:
        type: integer
      opration:
        $ref: '#/definitions/ds.Operation'
      quantity:
        type: integer
    type: object
...
  ds.Role:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - Buyer
    - Moderator
    - Admin
  ds.User:
    properties:
      balance:
        type: number
      login:
        type: string
      name:
        type: string
      password:
        type: string
      registrationDate:
        type: string
      role:
        $ref: '#/definitions/ds.Role'
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/operation-request-banknote:
    delete:
      consumes:
      - application/json
      description: Deletes a banknote from a request based on the user ID and banknote
        ID
      parameters:
      - description: banknote ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Delete banknote from request
      tags:
      - Operation_Banknote
    put:
      consumes:
      - application/json
      description: Update money Operation Banknote by client
      parameters:
      - description: Update quantity Operation Banknote
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.OperationBanknote'
      produces:
      - application/json
      responses:
        "200":
          description: update
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update money Operation Banknote
      tags:
      - Operation_Banknote
  /api/operations:
    delete:
      consumes:
      - application/json
      description: Deletes a operation request for the given user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Delete operation request by user ID
      tags:
      - Operations
    get:
      consumes:
      - application/json
      description: Retrieves a list of Operation requests based on the provided parameters
      parameters:
      - description: Operation request status
        in: query
        name: status_id
        type: string
      - description: Start date in the format '2006-01-02T15:04:05Z'
        in: query
        name: start_date
        type: string
      - description: End date in the format '2006-01-02T15:04:05Z'
        in: query
        name: end_date
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Operation'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get list of Operation requests
      tags:
      - Operations
    put:
      consumes:
      - application/json
      description: Update Operation by admin
      parameters:
      - description: updated Assembly
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Operation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update Operation by admin
      tags:
      - Operations
  /api/operations/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a operation request with the given ID
      parameters:
      - description: Operation Request ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
      summary: Get operation request by ID
      tags:
      - Operations
  /api/operations/form:
    put:
      consumes:
      - application/json
      description: Form Banknote by client
      parameters:
      - description: Operation form ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Form Banknote by client
      tags:
      - Operations
...
  /operations/updateStatus:
    put:
      consumes:
      - application/json
      description: Updates the status of a operation request with the given ID on
        "завершен"/"отклонен"
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: update status
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.NewStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Update operation request status by ID
      tags:
      - Operation
swagger: "2.0"